--- conpype/__init__.py
@@ -1,3 +1,0 @@
-from conpype.shell import shell, Password
-from conpype.concourse import *
-from conpype.resources import *
