--- conpype/concourse/__init__.py
@@ -1,3 +1,0 @@
-from conpype.concourse.pipeline import *
-from conpype.concourse.job import *
-from conpype.concourse.task import *
