+++ pipeline_dsl/__init__.py
@@ -1,0 +1,3 @@
+from pipeline_dsl.shell import shell, Password
+from pipeline_dsl.concourse import *
+from pipeline_dsl.resources import *
