+++ pipeline_dsl/resources/__init__.py
@@ -1,0 +1,12 @@
+from pipeline_dsl.resources.pool import Pool
+from pipeline_dsl.resources.gcs import GoogleCloudStorage
+from pipeline_dsl.resources.git import GitRepo
+from pipeline_dsl.resources.github import GithubRelease
+from pipeline_dsl.resources.docker import DockerImage
+from pipeline_dsl.resources.semver import SemVer, SemVerGitDriver
+from pipeline_dsl.resources.cron import Cron
+from pipeline_dsl.resources.registry_image import RegistryImage
+from pipeline_dsl.resources.github_pr import GithubPR
+
+ConcourseLockResource = Pool
+GoogleCloudStorageResource = GoogleCloudStorage
