+++ pipeline_dsl/resources/docker.py
@@ -1,8 +1,8 @@
 import os
-from conpype.concourse import concourse_context
+from pipeline_dsl.concourse import concourse_context
 
 
-class RegistryImageResource:
+class DockerImageResource:
     def __init__(self, name):
         self.name = name
         self.path = os.path.abspath(self.name)
@@ -10,20 +10,19 @@ class RegistryImageResource:
     def __str__(self):
         return self.name
 
-    def tag(self, default=None):
+    def digest(self):
         if concourse_context():
-            with open(os.path.join(self.path, "tag")) as f:
+            with open(os.path.join(self.path, "digest")) as f:
                 return f.read().strip()
-        return default
+        return "latest"
 
 
-class RegistryImage:
-    def __init__(self, repo, username=None, password=None, tag=None, variant=None):
-        self.repo = repo
+class DockerImage:
+    def __init__(self, repository, username=None, password=None, tag="latest"):
+        self.repository = repository
         self.username = username
         self.password = password
         self.tag = tag
-        self.variant = variant
 
     def resource_type(self):
         return None
@@ -31,19 +30,17 @@ class RegistryImage:
     def concourse(self, name):
         result = {
             "name": name,
-            "type": "registry-image",
+            "type": "docker-image",
+            "icon": "docker",
             "source": {
-                "repository": self.repo,
+                "repository": self.repository,
                 "username": self.username,
                 "password": self.password,
                 "tag": self.tag,
-                "variant": self.variant,
             },
         }
-        result["source"] = dict(
-            filter(lambda x: x[1] is not None, result["source"].items()),
-        )
+        result["source"] = dict(filter(lambda x: x[1] is not None, result["source"].items()))
         return result
 
     def get(self, name):
-        return RegistryImageResource(name)
+        return DockerImageResource(name)
