+++ pipeline_dsl/resources/github_pr.py
@@ -1,6 +1,6 @@
 import os
 import json
-from conpype.concourse import concourse_context
+from pipeline_dsl.concourse import concourse_context
 
 
 class GithubPRResource:
