+++ pipeline_dsl/resources/registry_image.py
@@ -1,8 +1,8 @@
 import os
-from conpype.concourse import concourse_context
+from pipeline_dsl.concourse import concourse_context
 
 
-class DockerImageResource:
+class RegistryImageResource:
     def __init__(self, name):
         self.name = name
         self.path = os.path.abspath(self.name)
@@ -10,19 +10,20 @@ class DockerImageResource:
     def __str__(self):
         return self.name
 
-    def digest(self):
+    def tag(self, default=None):
         if concourse_context():
-            with open(os.path.join(self.path, "digest")) as f:
+            with open(os.path.join(self.path, "tag")) as f:
                 return f.read().strip()
-        return "latest"
+        return default
 
 
-class DockerImage:
-    def __init__(self, repository, username=None, password=None, tag="latest"):
-        self.repository = repository
+class RegistryImage:
+    def __init__(self, repo, username=None, password=None, tag=None, variant=None):
+        self.repo = repo
         self.username = username
         self.password = password
         self.tag = tag
+        self.variant = variant
 
     def resource_type(self):
         return None
@@ -30,17 +31,19 @@ class DockerImage:
     def concourse(self, name):
         result = {
             "name": name,
-            "type": "docker-image",
-            "icon": "docker",
+            "type": "registry-image",
             "source": {
-                "repository": self.repository,
+                "repository": self.repo,
                 "username": self.username,
                 "password": self.password,
                 "tag": self.tag,
+                "variant": self.variant,
             },
         }
-        result["source"] = dict(filter(lambda x: x[1] is not None, result["source"].items()))
+        result["source"] = dict(
+            filter(lambda x: x[1] is not None, result["source"].items()),
+        )
         return result
 
     def get(self, name):
-        return DockerImageResource(name)
+        return RegistryImageResource(name)
