+++ pipeline_dsl/resources/semver.py
@@ -1,5 +1,5 @@
 import os
-from conpype.concourse import concourse_context
+from pipeline_dsl.concourse import concourse_context
 
 
 class SemVerResource:
