+++ pipeline_dsl/utils/stop-docker.sh
